Move Bundle extension to Extensions.swift
