+++ example/Auth/PhoneAuthView.swift
@@ -9,26 +9,6 @@ import SwiftUI
 import ParaSwift
 import Combine
 
-extension Bundle {
-    func decode<T: Decodable>(_ file: String) -> T {
-        guard let url = self.url(forResource: file, withExtension: nil) else {
-            fatalError("Failed to locate \(file) in bundle.")
-        }
-        
-        guard let data = try? Data(contentsOf: url) else {
-            fatalError("Failed to load \(file) from bundle.")
-        }
-        
-        let decoder = JSONDecoder()
-        
-        guard let loaded = try? decoder.decode(T.self, from: data) else {
-            fatalError("Failed to decode \(file) from bundle.")
-        }
-        
-        return loaded
-    }
-}
-
 struct CPData: Codable, Identifiable {
     let id: String
     let name: String
